Fix broken test, remove unused files
Removes a unit test for functions that no longer exist.

Also moves the last non-caching func from caching.go and removes the
file. Since atstccfg no longer does caching.
